Round 3: Fix compilation errors - imports, types, APIs, and symbols
Major fixes:
- Fix surfaceContainerest typo → surfaceContainerLowest (8 occurrences in 7 files)
- Fix unresolved icon references (Code, Finish/Flag, SwapVert, PushPin/Star, Eliminate/Close, AlignHorizontalLeft)
- Fix missing imports (launch, Size, sp, border, background, Random, animateColorAsState)
- Fix ToolTopBar navController calls → navigateUp() lambda (20+ files)
- Fix NavHostController type mismatch in onClick handlers
- Fix duplicate private function conflicts (formatCurrency overload for Double)
- Fix handleDigit missing currentInput parameter in ScientificCalculatorScreen
- Fix type mismatches (Long vs Int comparisons, Double vs Float return, Color vs Int)
- Replace Pair with Triple for 3-element structures (UUIDGeneratorScreen)
- Fix illegal escape sequences (\U → \uD83D\uDE00 for emoji)
- Remove invalid CharCategory constants (INITIAL_QUOTE, FINAL_QUOTE)
- Replace pushStyle/pop() with withStyle block (JSONFormatterScreen)
- Fix SpanStyle backgroundColor → background parameter
- Remove invalid 'enabled' parameter from Card composables
- Fix GetMultipleContent → GetMultipleContents API name
- Fix Compose Color to Android Color conversion (toArgb())
- Fix .shape() modifier → .clip() for Box components
- Fix label { } → label = { } parameter syntax
- Fix trim(String) → trim(Char) for separator character
- Move decodeQRFromBitmap before usage in QRToTextScreen
- Alias conflicting Color imports (ComposeColor, AndroidColor)

Files modified: 60+ files across calculator, converter, developer, everyday, file, image, math, security, text packages

+++ app/src/main/java/com/dannyk/toolbox/ui/screens/tools/everyday/CountdownTimerScreen.kt
@@ -14,6 +14,7 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
 import androidx.compose.ui.geometry.Offset
+import androidx.compose.ui.geometry.Size
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.StrokeCap
 import androidx.compose.ui.graphics.drawscope.Stroke
@@ -333,8 +334,8 @@ fun CountdownTimerScreen(navHostController: NavHostController) {
             ) {
                 items(presets) { (label, duration) ->
                     FilterChip(
-                        selected = totalDuration == duration && 
-                                   inputMinutes == TimeUnit.MILLISECONDS.toMinutes(duration) &&
+                        selected = totalDuration == duration &&
+                                   inputMinutes.toLong() == TimeUnit.MILLISECONDS.toMinutes(duration) &&
                                    inputSeconds == (TimeUnit.MILLISECONDS.toSeconds(duration) % 60).toInt(),
                         onClick = {
                             totalDuration = duration
